refactor: interface 활용하여 추후 확장될 공연 종류에 따른 서비스 확장성 개선
