+++ src/main/java/muse_kopis/muse/ticketbook/domain/photo/presentation/PhotoController.java
@@ -5,6 +5,7 @@ import java.util.ArrayList;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import muse_kopis.muse.ticketbook.domain.photo.application.PhotoService;
+import muse_kopis.muse.ticketbook.domain.photo.application.PhotoServiceImpl;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -33,7 +34,7 @@ public class PhotoController {
     @DeleteMapping
     @Operation(description = "사진 여러장을 한번에 지울 수 있습니다.(List<String>타입으로 되어있습니다.")
     public ResponseEntity<Void> deleteImage(@RequestBody List<String> photos) {
-        photos.forEach(photoService::deleteImageFromS3);
+        photos.forEach(photoService::deleteImage);
         return ResponseEntity.ok().build();
     }
 }
